MDL-68166 backup: Move backups to localcache
This set of changes is hard to break into separate commits.

In a typical backup, all files are stored, processed, and zipped in a
single HTTP/CLI request. As a result, it is safe to move them to a
localcache directory.

There are some situations where a backup _cannot_ simply move to
localcache, as there is an immedaite restore operation from the same
path. Those cases are:
- activity duplication
- course duplication
- course import

Unfortunately the following assumptions were originally made, and now
need to be broken:
- the backup directory is forever
- backup and restore are fundemantally identical they just have opposite
  steps
- the backup and restore processes use the same directory

In the case of an interactive restoration the directory _cannot_ be in
localcache because the same server may not be presented in a multi-node
environment. Therefore we must continue to use shared cache for the
restoration process for now.

Unfortunately it is not possible to apply the localcache directory rules
to a non-interactive restore as the process is divorced of information
about the restoration attempt beyond the backup id.