MDL-66979 behat: Work around geckodriver bug
A bug in geckodriver prevents the use of explicit NULL characters to
reset key state of modifier keys as per the W3C WebDriver specification.

See https://bugzilla.mozilla.org/show_bug.cgi?id=1494661 for details.

An implicit NULL is sent at the end of each set of keys to clear the
modifier key state and does work as expected.

This workaround takes the set of keys, explodes it on the NULL
character, and then calls sendKeys for each set of non-nulled keys
thereby achieving the same result.