MDL-49304 core: Ensure that LI are in an appropriate parent
All <li> tags must be in a valid parent. Valid parents are:
- <ul>
- <ol>
- <menu>
- <dir>

Where a list element is not found in a valid parent, it should be added
to an appropriate parent.

If the previous DOM Node sibling in the tree is a valid parent, then it
will be added to that parent.

If the previous DOM Node sibling is not a valid parent, then a new
parent will be created, with the default parent type being an unordered
list (UL).

Given the following example:

    <div>
      <p>First list</p>
      <li>Item one</li>
      <li>Item two</li>
    </div>
    <div>
      <p>Second list</p>
      <ul>
        <li>Item three</li>
      </ul>
      <li>Item four</li>
    </div>

The following would be the result:

    <div>
      <p>First list</p>
      <ul>
        <li>Item one</li>
        <li>Item two</li>
      </ul>
    </div>
    <div>
      <p>Second list</p>
      <ul>
        <li>Item three</li>
        <li>Item four</li>
      </ul>
    </div>

For `<li>Item one</li>` the previous sibling is `<p>First list</p>`. The
`<p>` tag is not a valid parent element to a `<li>` element. Therefore
one is created and the markup for that section becomes:

    <div>
      <p>First list</p>
      <ul>
        <li>Item one</li>
      </ul>
      <li>Item two</li>
    </div>

For `<li>Item two</li>` the previous sibling is now the newly created
`<ul>`, which is a valid parent to an `<li>` element.
Therefore the list element is moved into the unordered list, and the
markup for that section becomes: In the first case, there was no

    <div>
      <p>First list</p>
      <ul>
        <li>Item one</li>
        <li>Item two</li>
      </ul>
    </div>

The same happens for both Items three, and four.