MDL-49304 core_message: Parse text before passing to jQuery
jQuery can take an HTML element, such as the following, and create a
jQuery object from it:

    <p>Javascript is beautiful</p>

However, this only work where valid HTML is passed in.

The Messaging code assumes that the content passed to it is valid HTML,
but fails to consider that an HTMLText element is still valid HTML. That
is to say that the following is considered to be valid HTML:

    Javascript is beautiful

However, jQuery's object creator treats strings which do not look like
conventional HTML as a Search query filter which silently fails.

If the input string contains invalid markup, such as an exclamation
mark, this makes jQuery very unhappy. For example, the following will
lead to an error:

    Javascript is beautiful!

In order to create a jQuery object from this content, it must be valid
HTML. In this instance the object is only used to get a parsed textual
representation of the message content, so the following is safe to use:

    var messageContent document.createElement('p');
    messageContent.innerHTML = "Javascript is beautiful!"
    var messagePreview = $(messageContent).text();

This will correctly parse the HTML tag, and return the text such that
the result is "Javascript is beautiful!" once more.