MDL-69708 core: Introducing the new Content API
This new API introduces a standardised way for any part of Moodle,
including plugintypes, components (subsystems, and plugins) to content
in Moodle, how it is accessed, how it is served, and the checking of
access control on that content item.

Currently the API supports file content, but it will be extended to
support a standardised way to describe text content within Moodle too.