MDL-70137 cache: Reset cache configuration to default if invalid
When there are no stores, or no locks, regenerate the cache from empty.

There is no natural route to reach the sitaution where there are either
no cache locks defined, or no stores defined. That is to say that the UI
does not allow the default lock factory to be removed, nor any of the
default cache stores.

In addition to not allowing their removal, the UI will throw an
exception when trying to access any of the cache configuration changes
if there are no stores, or no lock systems. This means that it is not
possible to fix the configuration.

The most likely scenario here is that the cache has been corrupted
somehow and the safest thing to do is therefore to rewrite it.