MDL-70137 cache: Regenerate corrupt cache during write
Under certain situations we already detect a corrupt cache but we
previously warned about it without providing any way to fix it.

For this case, where we detect that no cache locks are defined, there is
no natural route to reach this situation - i.e. the UI does not allow
the default lock factory to be removed.

In addition the UI will throw an exception when trying to access any of
the cache configuration changes if there are no stores, or no lock
systems. This means that it is not possible to fix the configuration.

The most likely scenario here is that the cache has been corrupted
somehow and the safest thing to do is therefore to rewrite it.